优化；不能跑

+++ kernel/src/device/constant.rs
@@ -1,49 +1,49 @@
 /**
  * 数据寄存器的偏移量
  */
-pub const DATA_REGISTER_OFFSET: u8 = 0x0;
+pub const DATA_REGISTER_OFFSET: u16 = 0x0;
 
 /**
  * ERROR寄存器的偏移量
  */
-pub const ERROR_REGISTER_OFFSET: u8 = 0x1;
+pub const ERROR_REGISTER_OFFSET: u16 = 0x1;
 /**
  * Feature寄存器的偏移量
  */
-pub const FEATURE_REGISTER_OFFSET: u8 = 0x1;
+pub const FEATURE_REGISTER_OFFSET: u16 = 0x1;
 
 /**
  * Sector Count寄存器的偏移量
  */
-pub const SECTOR_COUNT_REGISTER_OFFSET: u8 = 0x2;
+pub const SECTOR_COUNT_REGISTER_OFFSET: u16 = 0x2;
 
 /**
  * LBA Low寄存器的偏移量
  */
-pub const LBA_LOW_REGISTER_OFFSET: u8 = 0x3;
+pub const LBA_LOW_REGISTER_OFFSET: u16 = 0x3;
 
 /**
  * LBA Mid寄存器的偏移量
  */
-pub const LBA_MID_REGISTER_OFFSET: u8 = 0x4;
+pub const LBA_MID_REGISTER_OFFSET: u16 = 0x4;
 
 /**
  * LBA Mid寄存器的偏移量
  */
-pub const LBA_HIGH_REGISTER_OFFSET: u8 = 0x5;
+pub const LBA_HIGH_REGISTER_OFFSET: u16 = 0x5;
 
 /**
  * device寄存器的偏移量
  */
-pub const DEVICE_REGISTER_OFFSET: u8 = 0x6;
+pub const DEVICE_REGISTER_OFFSET: u16 = 0x6;
 
 /**
  * status寄存器的偏移量
  */
-pub const STATUS_REGISTER_OFFSET: u8 = 0x7;
+pub const STATUS_REGISTER_OFFSET: u16 = 0x7;
 
 /**
  * command寄存器的偏移量
  */
-pub const COMMAND_REGISTER_OFFSET: u8 = 0x7;
+pub const COMMAND_REGISTER_OFFSET: u16 = 0x7;
